+++ app/utils/configuration.py
@@ -34,3 +34,11 @@ class ConfigurationManager:
     def get_hostname(self):
         """Get the hostname from the configuration"""
         return self.config.get("hostname")
+
+    def get_webcontrol_port(self):
+        """Get the port from the configuration"""
+        return self.config.get("control_server_port", 8080)
+
+    def get_webcontrol_host(self):
+        """Get the host from the configuration"""
+        return self.config.get("control_server_host", "127.0.0.1")
