+++ requirements.txt
@@ -1,3 +1,6 @@
 inotify_simple==1.3.5
 psutil==6.1.0
 sentry-sdk==2.17.0
+fastapi==0.115.3
+uvicorn==0.32.0
+python-multipart==0.0.16
