--- test/destination/source/stationary_test
@@ -1,1 +1,0 @@
-This is just to test if the app synced.
